Extract OverrideCompletion related logic from Completions.scala
